PYIC-1718: Rename the access-token lambda
