+++ lambdas/issue-client-access-token/src/main/resources/IpvLambdaJsonLayout.json
