+++ lambdas/issue-client-access-token/src/main/resources/log4j2.yaml
