+++ settings.gradle
@@ -1,7 +1,7 @@
 rootProject.name = 'di-ipv-core-back'
 include "lib",
 		"lambdas",
-		"lambdas:accesstoken",
+		"lambdas:issue-client-access-token",
 		"lambdas:sessionend",
 		"lambdas:credentialissuerreturn",
 		"lambdas:credentialissuerconfig",
